% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lininverse.R
\name{ce_lininv}
\alias{ce_lininv}
\title{Pure Inverse Problem}
\usage{
ce_lininv(y, X, p0, optim_method = "BFGS")
}
\arguments{
\item{y}{An (Nx1) vector representing the dependent variable where N is the number of observations}

\item{X}{An (NxK) matrix representing a set of independent variables where K is number of regressors}

\item{p0}{Optional: Prior probabilities associated with the regression coefficients}

\item{optim_method}{Optional: same as the "method" argument for the "optim" function in "stats"}
}
\value{
This function returns a list which has the following elements.
\itemize{
  \item lambda - Estimated Lagrange Multipliers.
  \item hess - Hessian matrix associated with the Lagrange Multipliers.
  \item p - Estimated probabilities associated with the regressions coefficients.
  \item p_e - Error bounds based on the weaker version of the Fano's inequality
  \item Sp - The (signal) information of the whole system.
  \item Pseudo_R2 - Pseudo R-squared.
  \item conv - convergence (same as in the optim function).
}
}
\description{
Pure Inverse Problem
}
\examples{
# Jaynes dice problem (Jaynes 1963)
y1 <- 2.5
y2 <- 3.5
y3 <- 4.5

X  <- matrix(1:6, nrow = 1)

p1 <- ce_lininv(y1, X)$p
p2 <- ce_lininv(y2, X)$p
p3 <- ce_lininv(y3, X)$p

par(mfrow=c(1,3))
barplot(p1, ylim = c(0, 0.4), names = X, main = paste0("Dice 1 (y = ",y1,")"), col = "red")
barplot(p2, ylim = c(0, 0.4), names = X, main = paste0("Dice 2 (y = ",y2,")"), col = "green")
barplot(p3, ylim = c(0, 0.4), names = X, main = paste0("Dice 3 (y = ",y3,")"), col = "blue")
}
