% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{gce_lin}
\alias{gce_lin}
\title{Linear Regression Model}
\usage{
gce_lin(y, X, Z, v, nu, p0, w0, optim_method = "BFGS")
}
\arguments{
\item{y}{An (Nx1) vector representing the dependent variable where N is the number of observations}

\item{X}{An (NxK) matrix representing a set of independent variables where K is number of regressors}

\item{Z}{An (KxM) matrix representing support spaces  the for regression coefficients where M is the dimension of the support spaces}

\item{v}{An optional argument representing a support space for error terms:
(a) if missing then \strong{v} is a (5x1) vector of equally spaced points in \code{[a,b]} interval;
(b) if a scalar (e.g. \strong{H}) then \strong{v} is a (Hx1) vector of equally spaced points in \code{[a,b]} interval;
(c) can be a user-supplied vector; (d) can be a user-supplied matrix.
Please note that in case (a) and (b) the \code{[a,b]} interval is centered around zero, and \strong{a} and \strong{b} are calculated
using the empirical three-sigma rule Pukelsheim (1994).}

\item{nu}{Optional: A weight parameter representing the trade-off between prediction and precision}

\item{p0}{Optional: Prior probabilities associated with the regression coefficients}

\item{w0}{Optional: Prior probabilities associated with the error terms}

\item{optim_method}{Optional: same as the "method" argument for the "optim" function in "stats"}
}
\value{
This function returns a list which has the following elements.
\itemize{
  \item lambda - Estimated Lagrange Multipliers.
  \item beta - Regression coefficients.
  \item var_beta - Variance-covariance matrix of the regression coefficients.
  \item p - Estimated probabilities associated with the regressions coefficients.
  \item w - Estimated probabilities associated with the error terms.
  \item e - Estimated Residuals.
  \item Sp - The (signal) information of the whole system.
  \item Sp_k - The (signal) information associated with the k-th regression coefficient.
  \item H_p_w - Value of the joint entropies of \code{p} and \code{w} at the final iteration.
  \item dH - Delta-H from the Entropy Concentration Theorem.
  \item ER - Entropy Ratio Statistic.
  \item Pseudo_R2 - Pseudo R-squared.
  \item conv - convergence (same as in the optim function).
}
}
\description{
This function estimates the General Linear Model (GLM) using the infometrics framework.
             The function provides possibilities to incorporate prior information, adjust weights on
             precision and prediction. Output of the function contains statistics necessary for
             inference and diagnostics. Further details can be found in Golan, Judge and Miller
             (1996, Chapter 6), Golan (2008, Chapter 6)
}
\examples{
set.seed(123)
y <- runif(100)
X <- matrix(runif(200), nrow = 100, ncol = 2)
Z <- matrix(c(-1, 0, 1, -1, 0, 1), nrow = 2, byrow = TRUE)
gce_lin(y, X, Z)
gce_lin(y, X, Z, 3)
gce_lin(y, X, Z, 3, 0.6)
}
