% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov.R
\name{gce_matrix}
\alias{gce_matrix}
\title{Matrix Balancing}
\usage{
gce_matrix(y, x, dimV, nu, p0, w0, optim_method = "BFGS")
}
\arguments{
\item{y}{A vector representing the sum of all the cells in respective rows of the unknown P matrix}

\item{x}{A vector representing the sum of all the cells in respective columns of the unknown P matrix}

\item{dimV}{An optional argument (scalar) representing dimensions of the support space for the error terms, default value is 5.}

\item{nu}{An optional argument (scalar in \code{[0,1]}) representing the trade-off between prediction and precision}

\item{p0}{Optional: Prior probabilities associated with the P matrix}

\item{w0}{Optional: Prior probabilities associated with the error terms}

\item{optim_method}{Optional: same as the "method" argument for the "optim" function in "stats"}
}
\value{
This function returns a list which has the following elements.
\itemize{
  \item lambda - Estimated Lagrange Multipliers.
  \item hess - Hessian matrix associated with the Lagrange Multipliers.
  \item p - Estimated transition matrix.
  \item w - Estimated probabilities associated with the error terms.
  \item e - Estimated Residuals.
  \item Sp - The (signal) information of the whole system.
  \item S_p_j - Reflects the (signal) information in each moment.
  \item p_e_j - Error bounds based on the weaker version of the Fano's inequality
  \item H_p_w - Value of the joint entropies of \code{p} and \code{w} at the final iteration.
  \item ER - Entropy Ratio Statistic.
  \item Pseudo_R2 - Pseudo R-squared.
  \item conv - convergence (same as in the optim function).
}
}
\description{
This function
}
\examples{
y <- c(0.3, 0.2, 0.5)
x <- c(0.1, 0.25, 0.3, 0.2, 0.15)
gce_matrix(y, x)
gce_matrix(y, x, 3)
gce_matrix(y, x, 3, 0.6)
}
