% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.R
\name{gme_mixed}
\alias{gme_mixed}
\title{Mixed Discrete Choice Model}
\usage{
gme_mixed(Y, X1, X2, dimS, dimV, optim_method = "BFGS")
}
\arguments{
\item{Y}{Dependent Variable: an (NxJ) matrix where N represents the number of individuals and J is the number of choices}

\item{X1}{Alternative-invariant regressors: (NxK1) matrix where K1 represents the number of regressors}

\item{X2}{Alternative-variant regressors: (NxJxK2) matrix where K2 represents the number of regressors}

\item{dimS}{An optional argument (scalar) representing dimensions of the support space for the phi probabilities, default value is 3}

\item{dimV}{An optional argument (scalar) representing dimensions of the support space for the error terms, default value is 5}

\item{optim_method}{Optional: same as the "method" argument for the "optim" function in "stats"}
}
\value{
A list which has the following elements:
\itemize{
  \item lambda - Lagrange multipliers associated with the alternative-invariant regressors.
  \item alpha - Lagrange multipliers associated with the alternative-variant regressors.
  \item rho - Lagrange multipliers associated with the normalization constraint.
  \item hess - Hessian matrix associated with the estimated Lagrange multipliers.
  \item H_phi_w - Value of the joint entropies of \code{phi} and \code{w} at the final iteration.
  \item marg_eff_x1 - Marginal effects associated with the alternative-invariant regressors.
  \item marg_eff_x2 - Marginal effects associated with the alternative-variant regressors.
  \item S1 - The (signal) information of the whole system.
  \item S2 - The (signal) information associated with the choice \code{j} of the individual \code{i}.
  \item Pseudo_R2 - Pseudo R-squared.
  \item ER - Entropy Ratio Statistic.
  \item conv - convergence (same as in the optim function).
}
}
\description{
Mixed Discrete Choice Model
}
\examples{
Y  <- c(1, 0, 0,
        0, 1, 0,
        0, 1, 0,
        1, 0, 0,
        0, 0, 1)
Y  <- matrix(Y, nrow = 5, byrow = TRUE)
X1 <- c(1, 2,
        1, 3,
        1, 1,
        1, 5,
        1, 4)
X1 <-matrix(X1, nrow = 5, byrow = TRUE)
X2 <- runif(30)
X2 <- array(X2, c(5, 3, 2))
gme_mixed(Y, X1, X2)
}
