% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lininverse.R
\name{lin_inv}
\alias{lin_inv}
\title{Linear Inverse Problem with Noise}
\usage{
lin_inv(y, X, v, nu, p0, w0, optim_method = "BFGS")
}
\arguments{
\item{y}{An (Nx1) vector representing the dependent variable where N is the number of observations}

\item{X}{An (NxK) matrix representing a set of independent variables where K is number of regressors}

\item{v}{An optional argument representing a support space for error terms:
(a) if missing then \strong{v} is a (5x1) vector of equally spaced points in \code{[a,b]} interval;
(b) if a scalar (e.g. \strong{H}) then \strong{v} is a (Hx1) vector of equally spaced points in \code{[a,b]} interval;
(c) can be a user-supplied vector; (d) can be a user-supplied matrix.
Please note that in case (a) and (b) the \code{[a,b]} interval is centered around zero, and \strong{a} and \strong{b} are calculated
using the empirical three-sigma rule Pukelsheim (1994).}

\item{nu}{Optional: A weight parameter representing the trade-off between prediction and precision}

\item{p0}{Optional: Prior probabilities associated with the regression coefficients}

\item{w0}{Optional: Prior probabilities associated with the error terms}

\item{optim_method}{Optional: same as the "method" argument for the "optim" function in "stats"}
}
\value{
This function returns a list which has the following elements.
\itemize{
  \item lambda - Estimated Lagrange Multipliers.
  \item hess - Hessian matrix associated with the Lagrange Multipliers.
  \item p - Estimated probabilities associated with the regressions coefficients.
  \item p_e - Error bounds based on the weaker version of the Fano's inequality
  \item w - Estimated probabilities associated with the error terms.
  \item e - Estimated Residuals.
  \item Sp - The (signal) information of the whole system.
  \item H_p_w - Value of the joint entropies of \code{p} and \code{w} at the final iteration.
  \item ER - Entropy Ratio Statistic.
  \item Pseudo_R2 - Pseudo R-squared.
  \item conv - convergence (same as in the optim function).
}
}
\description{
Linear Inverse Problem with Noise
}
\examples{
set.seed(123)
y <- runif(100)
X <- matrix(runif(200), nrow = 100, ncol = 2)
Z <- matrix(c(-1, 0, 1, -1, 0, 1), nrow = 2, byrow = TRUE)
gce_lin(y, X, Z)
gce_lin(y, X, Z, 3)
gce_lin(y, X, Z, 3, 0.6)
}
